requirements/misc.txt
#!/usr/bin/env -S pip install -r

tree-sitter-lsp[misc]
